Put to int

+++ experiments/scripts/finetune_independently.py
@@ -39,8 +39,8 @@ def initialize_nets(obj_detect_weights):
 def do_finetuning(id, finetuning_config, plotter, box_head_classification, box_predictor_classification):
     dataset = pickle.load(open("training_set/feature_training_set_track_{}.pkl".format(id), "rb"))
     #dataset.clean()
-    train_size =  0.8 * len(dataset)
-    test_size = len(dataset) - train_size
+    train_size =  int(0.8 * len(dataset))
+    test_size = int(len(dataset) - train_size)
     training_set, validation_set = torch.utils.data.random_split(dataset, [train_size, test_size])
 
     box_predictor_classification.train()
